fix: extract resolveProviderAlias into pure module to fix client build
The 'use client' component ComboControlCenterClient.tsx imports
controlCenter.ts which imported resolveProviderAlias from
open-sse/services/model.ts. That module has dynamic DB imports
(activeSyncedCatalog, readCache) that Turbopack bundles into the
client chunk, pulling in the playwright/sharp chain and causing 168
'Module not found' errors.

Extracted resolveProviderAlias and the alias maps into a new pure
module open-sse/config/providerAlias.ts that has no DB dependencies.
Updated controlCenter.ts to import from the pure module. model.ts
now re-exports from the pure module to avoid code duplication.

Fixes #13474
