+++ package.json
@@ -456,7 +456,12 @@
       "esbuild",
       "omniroute",
       "sharp"
-    ]
+    ],
+    "overrides": {
+      "lodash-es@>=4.0.0 <=4.17.23": ">=4.18.0",
+      "brace-expansion@>=4.0.0 <5.0.8": ">=5.0.9",
+      "brace-expansion@>=4.0.0 <5.0.9": ">=5.0.9"
+    }
   },
   "allowScripts": {
     "better-sqlite3": true,
